Learning how to use gir
Using gudev-rs as an example, this creates a ostree-sys to wrap the low
level ffi calls, Run make gir in the ./ostree-sys/ directory to
regenerate.

Top level autogenerated code is created with make gir at the top.
There are a large number of warnings from gir about 'has unsupported
outs' and an error on ChecksumInputStream.

src/lib.rs is hand-written.